// src/api/types.gen.ts
export type Aanhaling = {
  readonly nadrukProperty?: Array<Nadruk> | null;
  readonly nadrukPropertySpecified?: boolean;
  readonly sup?: Array<string> | null;
  readonly supSpecified?: boolean;
  readonly inf?: Array<string> | null;
  readonly infSpecified?: boolean;
  readonly ovl?: Array<string> | null;
  readonly ovlSpecified?: boolean;
  readonly unl?: Array<string> | null;
  readonly unlSpecified?: boolean;
  readonly omissie?: Array<Omissie> | null;
  readonly omissieSpecified?: boolean;
  bron?: string | null;
  url?: string | null;
  auteur?: string | null;
  lang?: Lang;
  langSpecified?: boolean;
  text?: Array<string> | null;
};
// ...
export type Al = {
  readonly nadrukProperty?: Array<Nadruk> | null;
  readonly nadrukPropertySpecified?: boolean;
  readonly sup?: Array<string> | null;
  readonly supSpecified?: boolean;
  readonly inf?: Array<string> | null;
  readonly infSpecified?: boolean;
  readonly ovl?: Array<string> | null;
  readonly ovlSpecified?: boolean;
  readonly unl?: Array<string> | null;
  readonly unlSpecified?: boolean;
  readonly omissie?: Array<Omissie> | null;
  readonly omissieSpecified?: boolean;
  readonly nootref?: Array<Nootref> | null;
  readonly nootrefSpecified?: boolean;
  readonly noot?: Array<Noot> | null;
  readonly nootSpecified?: boolean;
  readonly aanhaling?: Array<Aanhaling> | null;
  readonly aanhalingSpecified?: boolean;
  readonly intref?: Array<Intref> | null;
  readonly intrefSpecified?: boolean;
  readonly extref?: Array<Extref> | null;
  readonly extrefSpecified?: boolean;
  readonly extrefGroep?: Array<ExtrefGroep> | null;
  readonly extrefGroepSpecified?: boolean;
  readonly intrefGroep?: Array<IntrefGroep> | null;
  readonly intrefGroepSpecified?: boolean;
  readonly dossierref?: Array<Dossierref> | null;
  readonly dossierrefSpecified?: boolean;
  readonly indexterm?: Array<Indexterm> | null;
  readonly indextermSpecified?: boolean;
  readonly afk?: Array<Afk> | null;
  readonly afkSpecified?: boolean;
  readonly cit?: Array<Cit> | null;
  readonly citSpecified?: boolean;
  readonly dfn?: Array<Dfn> | null;
  readonly dfnSpecified?: boolean;
  readonly redactie?: Array<Redactie> | null;
  readonly redactieSpecified?: boolean;
  readonly tekstcorrectie?: Array<Tekstcorrectie> | null;
  readonly tekstcorrectieSpecified?: boolean;
  text?: Array<string> | null;
  id?: string | null;
  lang?: Lang;
  langSpecified?: boolean;
};
// ...
export type Artikel = {
  kop: Kop;
  readonly lid?: Array<Lid> | null;
  readonly lidSpecified?: boolean;
  readonly structuurAlgemeen?: Array<unknown> | null;
  readonly structuurAlgemeenSpecified?: boolean;
  readonly table?: Array<Table> | null;
  readonly tableSpecified?: boolean;
  readonly wetcitaat?: Array<Wetcitaat> | null;
  readonly wetcitaatSpecified?: boolean;
  readonly binnenwet?: Array<Binnenwet> | null;
  readonly binnenwetSpecified?: boolean;
  artikelToelichting?: ArtikelToelichting;
  metaData?: MetaData;
  id?: string | null;
  lang?: Lang;
  langSpecified?: boolean;
  status?: ILokaalAttArtikelStatus;
  code?: string | null;
  type?: string | null;
  labelId?: string | null;
  stamId?: string | null;
  versieId?: string | null;
  publicatie?: IBwbcmsAttributenPublicatie;
  publicatieSpecified?: boolean;
  terugwerking?: string | null;
  variabelDeel?: string | null;
};

export type ArtikelToelichting = {
  readonly structuurAlgemeen?: Array<unknown> | null;
  readonly structuurAlgemeenSpecified?: boolean;
  readonly table?: Array<Table> | null;
  readonly tableSpecified?: boolean;
  id?: string | null;
  lang?: Lang;
  langSpecified?: boolean;
};
// ...
export type Dossierref = {
  readonly nadrukProperty?: Array<Nadruk> | null;
  readonly nadrukPropertySpecified?: boolean;
  readonly sup?: Array<string> | null;
  readonly supSpecified?: boolean;
  readonly inf?: Array<string> | null;
  readonly infSpecified?: boolean;
  readonly ovl?: Array<string> | null;
  readonly ovlSpecified?: boolean;
  readonly unl?: Array<string> | null;
  readonly unlSpecified?: boolean;
  readonly omissie?: Array<Omissie> | null;
  readonly omissieSpecified?: boolean;
  readonly nootref?: Array<Nootref> | null;
  readonly nootrefSpecified?: boolean;
  readonly noot?: Array<Noot> | null;
  readonly nootSpecified?: boolean;
  readonly aanhaling?: Array<Aanhaling> | null;
  readonly aanhalingSpecified?: boolean;
  dossier: string;
  text?: Array<string> | null;
};
// ...
export type Extref = {
  readonly nadrukProperty?: Array<Nadruk> | null;
  readonly nadrukPropertySpecified?: boolean;
  readonly sup?: Array<string> | null;
  readonly supSpecified?: boolean;
  readonly inf?: Array<string> | null;
  readonly infSpecified?: boolean;
  readonly ovl?: Array<string> | null;
  readonly ovlSpecified?: boolean;
  readonly unl?: Array<string> | null;
  readonly unlSpecified?: boolean;
  readonly omissie?: Array<Omissie> | null;
  readonly omissieSpecified?: boolean;
  readonly nootref?: Array<Nootref> | null;
  readonly nootrefSpecified?: boolean;
  readonly noot?: Array<Noot> | null;
  readonly nootSpecified?: boolean;
  readonly aanhaling?: Array<Aanhaling> | null;
  readonly aanhalingSpecified?: boolean;
  doc: string;
  struct?: string | null;
  anker?: string | null;
  reeks?: Reeks;
  reeksSpecified?: boolean;
  rol?: string | null;
  beschrijving?: string | null;
  label?: string | null;
  status?: IAttExtrefStatus;
  text?: Array<string> | null;
};

export type ExtrefGroep = {
  readonly nadrukProperty?: Array<Nadruk> | null;
  readonly nadrukPropertySpecified?: boolean;
  readonly sup?: Array<string> | null;
  readonly supSpecified?: boolean;
  readonly inf?: Array<string> | null;
  readonly infSpecified?: boolean;
  readonly ovl?: Array<string> | null;
  readonly ovlSpecified?: boolean;
  readonly unl?: Array<string> | null;
  readonly unlSpecified?: boolean;
  readonly omissie?: Array<Omissie> | null;
  readonly omissieSpecified?: boolean;
  readonly nootref?: Array<Nootref> | null;
  readonly nootrefSpecified?: boolean;
  readonly noot?: Array<Noot> | null;
  readonly nootSpecified?: boolean;
  readonly aanhaling?: Array<Aanhaling> | null;
  readonly aanhalingSpecified?: boolean;
  readonly refItem?: Array<RefItem> | null;
  readonly refItemSpecified?: boolean;
  readonly extref?: Array<Extref> | null;
  readonly extrefSpecified?: boolean;
  text?: Array<string> | null;
};
// ...
export type Hoofdstuk = {
  kop: Kop;
  structuurtekst?: Structuurtekst;
  readonly titeldeel?: Array<Titeldeel> | null;
  readonly titeldeelSpecified?: boolean;
  readonly afdeling?: Array<Afdeling> | null;
  readonly afdelingSpecified?: boolean;
  readonly paragraaf?: Array<Paragraaf> | null;
  readonly paragraafSpecified?: boolean;
  readonly artikel?: Array<Artikel> | null;
  readonly artikelSpecified?: boolean;
  readonly lokaalStructuurWetgeving?: Array<ClassWijzigArtikel> | null;
  readonly lokaalStructuurWetgevingSpecified?: boolean;
  metaData?: MetaData;
  id?: string | null;
  status?: IAttlistHoofdstukStatus;
  statusSpecified?: boolean;
  terugwerking?: string | null;
  labelId?: string | null;
  stamId?: string | null;
  versieId?: string | null;
  publicatie?: IBwbcmsAttributenPublicatie;
  publicatieSpecified?: boolean;
};
// ...
export type ILokaalAttArtikelStatus = 0 | 1 | 2;

export type ILokaalAttBijlageStatus = 0 | 1 | 2;

export type ILokaalAttInhoudStatus = 0 | 1 | 2;

export type ILokaalAttNrStatus = 0 | 1;

export type ILokaalAttTitelStatus = 0 | 1;

export type ILokaalAttWetsluitingStatus = 0 | 1 | 2;

export type Indexterm = {
  readonly nadrukProperty?: Array<Nadruk> | null;
  readonly nadrukPropertySpecified?: boolean;
  readonly sup?: Array<string> | null;
  readonly supSpecified?: boolean;
  readonly inf?: Array<string> | null;
  readonly infSpecified?: boolean;
  readonly ovl?: Array<string> | null;
  readonly ovlSpecified?: boolean;
  readonly unl?: Array<string> | null;
  readonly unlSpecified?: boolean;
  readonly omissie?: Array<Omissie> | null;
  readonly omissieSpecified?: boolean;
  readonly nootref?: Array<Nootref> | null;
  readonly nootrefSpecified?: boolean;
  readonly noot?: Array<Noot> | null;
  readonly nootSpecified?: boolean;
  readonly aanhaling?: Array<Aanhaling> | null;
  readonly aanhalingSpecified?: boolean;
  id?: string | null;
  term?: string | null;
  indexgroep?: string | null;
  text?: Array<string> | null;
};
// ...
export type Intref = {
  readonly nadrukProperty?: Array<Nadruk> | null;
  readonly nadrukPropertySpecified?: boolean;
  readonly sup?: Array<string> | null;
  readonly supSpecified?: boolean;
  readonly inf?: Array<string> | null;
  readonly infSpecified?: boolean;
  readonly ovl?: Array<string> | null;
  readonly ovlSpecified?: boolean;
  readonly unl?: Array<string> | null;
  readonly unlSpecified?: boolean;
  readonly omissie?: Array<Omissie> | null;
  readonly omissieSpecified?: boolean;
  readonly nootref?: Array<Nootref> | null;
  readonly nootrefSpecified?: boolean;
  readonly noot?: Array<Noot> | null;
  readonly nootSpecified?: boolean;
  readonly aanhaling?: Array<Aanhaling> | null;
  readonly aanhalingSpecified?: boolean;
  doc?: string | null;
  struct?: string | null;
  anker?: string | null;
  reeks?: Reeks;
  reeksSpecified?: boolean;
  rol?: string | null;
  beschrijving?: string | null;
  label?: string | null;
  status?: IAttIntrefStatus;
  text?: Array<string> | null;
};

export type IntrefGroep = {
  readonly nadrukProperty?: Array<Nadruk> | null;
  readonly nadrukPropertySpecified?: boolean;
  readonly sup?: Array<string> | null;
  readonly supSpecified?: boolean;
  readonly inf?: Array<string> | null;
  readonly infSpecified?: boolean;
  readonly ovl?: Array<string> | null;
  readonly ovlSpecified?: boolean;
  readonly unl?: Array<string> | null;
  readonly unlSpecified?: boolean;
  readonly omissie?: Array<Omissie> | null;
  readonly omissieSpecified?: boolean;
  readonly nootref?: Array<Nootref> | null;
  readonly nootrefSpecified?: boolean;
  readonly noot?: Array<Noot> | null;
  readonly nootSpecified?: boolean;
  readonly aanhaling?: Array<Aanhaling> | null;
  readonly aanhalingSpecified?: boolean;
  readonly refItem?: Array<RefItem> | null;
  readonly refItemSpecified?: boolean;
  readonly intref?: Array<Intref> | null;
  readonly intrefSpecified?: boolean;
  text?: Array<string> | null;
};
// ...
export type Lid = {
  lidnr: Lidnr;
  readonly structuurAlgemeen?: Array<unknown> | null;
  readonly structuurAlgemeenSpecified?: boolean;
  readonly table?: Array<Table> | null;
  readonly tableSpecified?: boolean;
  readonly wetcitaat?: Array<Wetcitaat> | null;
  readonly wetcitaatSpecified?: boolean;
  readonly binnenwet?: Array<Binnenwet> | null;
  readonly binnenwetSpecified?: boolean;
  id?: string | null;
  lang?: Lang;
  langSpecified?: boolean;
};
// ...
export type Nadruk = {
  readonly nadrukProperty?: Array<Nadruk> | null;
  readonly nadrukPropertySpecified?: boolean;
  readonly sup?: Array<string> | null;
  readonly supSpecified?: boolean;
  readonly inf?: Array<string> | null;
  readonly infSpecified?: boolean;
  readonly ovl?: Array<string> | null;
  readonly ovlSpecified?: boolean;
  readonly unl?: Array<string> | null;
  readonly unlSpecified?: boolean;
  readonly omissie?: Array<Omissie> | null;
  readonly omissieSpecified?: boolean;
  readonly nootref?: Array<Nootref> | null;
  readonly nootrefSpecified?: boolean;
  readonly noot?: Array<Noot> | null;
  readonly nootSpecified?: boolean;
  readonly aanhaling?: Array<Aanhaling> | null;
  readonly aanhalingSpecified?: boolean;
  type?: IAttlistNadrukType;
  text?: Array<string> | null;
};

export type Noot = {
  nootNr?: string | null;
  readonly nootAl?: Array<ClassNootAl> | null;
  readonly nootAlSpecified?: boolean;
  readonly naamlijst?: Array<Naamlijst> | null;
  readonly naamlijstSpecified?: boolean;
  readonly nootLijst?: Array<NootLijst> | null;
  readonly nootLijstSpecified?: boolean;
  id?: string | null;
  type?: IAttNootTypeType;
  lang?: Lang;
  langSpecified?: boolean;
};
// ...
export type Nootref = {
  refid: string;
};
// ...
export type Omissie = {
  reden?: string | null;
  bron?: string | null;
  plaatsing?: IAttlistOmissiePlaatsing;
};
// ...
export type Tekstcorrectie = {
  id?: string | null;
  text?: Array<string> | null;
};